+++ app/src/main/AndroidManifest.xml
@@ -17,10 +17,10 @@
         android:theme="@style/Theme.CalificacionesUNPA"
         tools:targetApi="31">
         <activity
-            android:name=".notificaciones_escolares"
+            android:name=".ui.login.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.login.LoginActivity"
+            android:name=".notificaciones_escolares"
             android:exported="true"
             android:label="CalificaionesUNPA"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
