+++ app/src/main/java/mx/edu/unpa/calificacionesunpa/adapters/NotificacionAdapter.kt
@@ -13,7 +13,8 @@ data class NotificationItem(
     val titulo: String = "",
     val mensaje: String = "",
     val esGlobal: Boolean = false,
-    val timestamp: Long = 0L
+    val timestamp: Long = 0L,
+    val fueLeida : Boolean = true
 ) {
     constructor() : this(0, "", "", false, 0L) // Constructor sin argumentos requerido por Firebase
 }
