+++ app/src/main/java/mx/edu/unpa/calificacionesunpa/service/NotificacionesService.kt
@@ -1,0 +1,7 @@
+package mx.edu.unpa.calificacionesunpa.service
+
+import mx.edu.unpa.calificacionesunpa.adapters.NotificationItem
+
+object NotificacionesService {
+    var listaNotificaciones : List<NotificationItem> = emptyList()
+}
