+++ app/src/main/res/layout/notificaciones_escolares.xml
@@ -28,7 +28,7 @@
                 android:text="Notificaciones"
                 android:textSize="50sp"
                 android:textStyle="bold|italic"
-                android:textColor="#8B0000"/>
+                android:textColor="#000000"/>
 
             <!-- Pregunta -->
             <TextView
@@ -40,7 +40,7 @@
                 android:text="¿A quién desea enviar la notificación?"
                 android:textSize="23sp"
                 android:textStyle="bold|italic"
-                android:textColor="#00008B"/>
+                android:textColor="#000000"/>
 
             <!-- Radios -->
             <RadioGroup
@@ -61,7 +61,7 @@
                 <RadioButton
                     android:id="@+id/radioCiertos"
                     android:layout_width="156dp"
-                    android:layout_height="48dp"
+                    android:layout_height="match_parent"
                     android:text="Ciertos alumnos"
                     android:textColor="#000000"
                     android:textSize="15sp"
@@ -109,21 +109,21 @@
                 android:visibility="gone"
                 android:layout_marginTop="12dp">
 
-                <TextView
+                <!--<TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="Selecciona la carrera" />
+                    android:text="Selecciona la carrera" />-->
 
                 <Spinner
                     android:id="@+id/spinnerCarrera"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content" />
 
-                <TextView
+                <!--<TextView
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:text="Selecciona el semestre"
-                    android:layout_marginTop="8dp" />
+                    android:layout_marginTop="8dp" />-->
 
                 <Spinner
                     android:id="@+id/spinnerSemestre"
